Embed QuestReward in quest. Add reward generators
